Reorganized types and interfaces
