+++ src/types/index.ts
@@ -1,10 +1,6 @@
 // Entity typing reference: https://community.onespan.com/sites/default/files/interactive_api_references/oss_1145.yaml
 
-export * from './package';
-export * from './document';
-export * from './external';
-export * from './message';
-export * from './role';
-export * from './translation';
-export * from './user';
+export * from './models';
+export * as Requests from './requests';
+export * as Responses from './responses';
 export * from './utils';
