+++ src/types/models/external.ts
@@ -1,4 +1,4 @@
-import { Nullable } from './utils';
+import { Nullable } from '../utils';
 
 /** @public */
 export interface External {
