+++ src/types/models/translation.ts
