+++ src/types/responses/package.ts
@@ -1,0 +1,4 @@
+/** @public */
+export interface CreatePackage {
+  id: string;
+}
